Added Swagger docs for Task endpoints

+++ src/main/java/app/tasks/handler/TaskHandler.java
@@ -8,6 +8,9 @@ import app.tasks.repository.TaskRepository;
 import app.tasks.service.QueryService;
 import app.tasks.utils.AuthUtils;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,25 +40,42 @@ public class TaskHandler {
         this.shareRepository = shareRepository;
     }
 
-    @Operation(security = {@SecurityRequirement(name = "Authorization")})
+    @Operation(summary = "Get all tasks of a user with any level of access. Returns list of task and access attributes", security = {@SecurityRequirement(name = "Authorization")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Get list of tasks and access attributes. Empty list if no tasks",content = @Content)
+    })
     @GetMapping(value = "/tasks")
     public List<Map<String, Object>> getTask(@RequestHeader("Authorization") String sessionToken) {
         String userId = authUtils.isAuthenticated(sessionToken, sessionRepository);
         return queryService.executeQueryResponse(GET_TASKS_WITH_ACCESS, Map.of("userId", userId));
     }
 
-    @Operation(security = {@SecurityRequirement(name = "Authorization")})
+    @Operation(summary = "Get a specific task with taskId if user has access",security = {@SecurityRequirement(name = "Authorization")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Get tasks and access attributes of specific task",  content = @Content),
+            @ApiResponse(responseCode = "404", description = "Fails if task doesn't exist or user doesn't have access",  content = @Content)
+    })
     @GetMapping(value = "/task/{taskId}")
     public List<Map<String, Object>> getTaskByTaskId(@PathVariable String taskId, @RequestHeader("Authorization") String sessionToken) {
         String userId = authUtils.isAuthenticated(sessionToken, sessionRepository);
-        return queryService.executeQueryResponse(GET_TASK_WITH_ACCESS, Map.of("userId", userId, "taskId", taskId));
+        List<Map<String, Object>> result = queryService.executeQueryResponse(GET_TASK_WITH_ACCESS, Map.of("userId", userId, "taskId", taskId));
+        if(result.size()>0){
+            return result;
+        }
+        else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No such task");
+        }
     }
 
 
-    @Operation(security = {@SecurityRequirement(name = "Authorization")})
+    @Operation(summary = "Create task. User is automatically granted edit access to the task", security = {@SecurityRequirement(name = "Authorization")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Returns JSON containing of newly created taskId with key 'id' ",content = @Content)
+    })
     @PostMapping(value = "/task", consumes = MediaType.APPLICATION_JSON_VALUE)
     @Transactional
-    public Map<String, String> createTask(@RequestBody Task taskInput, @RequestHeader("Authorization") String sessionToken) {
+    public Map<String, String> createTask(@io.swagger.v3.oas.annotations.parameters.RequestBody(description = "id, userId and lastUpdateTs will be overwritten by server")
+                                              @RequestBody Task taskInput, @RequestHeader("Authorization") String sessionToken) {
         String userId = authUtils.isAuthenticated(sessionToken, sessionRepository);
         String taskId = UUID.randomUUID().toString();
 
@@ -70,10 +90,15 @@ public class TaskHandler {
         return Map.of("id", taskInput.getId());
     }
 
-    @Operation(security = {@SecurityRequirement(name = "Authorization")})
+    @Operation(summary = "To update a task. Returns 404 if task not found or no access", security = {@SecurityRequirement(name = "Authorization")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "No return value",content = @Content),
+            @ApiResponse(responseCode = "404", description = "Task not found or no access",content = @Content)
+    })
     @PutMapping(value = "/task", consumes = MediaType.APPLICATION_JSON_VALUE)
     @Transactional
-    public void updateTask(@RequestBody Task taskInput, @RequestHeader("Authorization") String sessionToken) {
+    public void updateTask(@io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Only id(taskId) is required. Any other field provided will updated if it matches type validation")
+                               @RequestBody Task taskInput, @RequestHeader("Authorization") String sessionToken) {
         String userId = authUtils.isAuthenticated(sessionToken, sessionRepository);
         List<Task> taskObj = taskRepository.getTaskByIdAndUserIdAndCheckAccess(taskInput.getId(),userId);
         if(taskObj.size()==1){
@@ -93,12 +118,15 @@ public class TaskHandler {
         }
     }
 
-
-
-    @Operation(security = {@SecurityRequirement(name = "Authorization")})
+    @Operation(summary = "To delete a list of tasks. Tasks with edit access for the user are deleted. List of deleted tasks is returned", security = {@SecurityRequirement(name = "Authorization")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "List of deleted tasks is returned",content = @Content),
+            @ApiResponse(responseCode = "400", description = "No tasks deleted",content = @Content)
+    })
     @DeleteMapping(value = "/tasks", consumes = MediaType.APPLICATION_JSON_VALUE)
     @Transactional
-    public void deleteTask(@RequestBody Collection<String> taskIds, @RequestHeader("Authorization") String sessionToken) {
+    public List<String> deleteTask(@io.swagger.v3.oas.annotations.parameters.RequestBody(description = "List of taskIds to be deleted. User must have edit access to the tasks")
+                                       @RequestBody Collection<String> taskIds, @RequestHeader("Authorization") String sessionToken) {
         String userId = authUtils.isAuthenticated(sessionToken, sessionRepository);
         List<ShareModel> shares = shareRepository.findByTaskIdInAndUserId(taskIds, userId);
 
@@ -107,8 +135,12 @@ public class TaskHandler {
             if (Objects.equals(s.getAccessType(), "edit"))
                 tasksWithAccess.add(s.getTaskId());
         }
-        taskRepository.deleteAllById(tasksWithAccess);
-        shareRepository.deleteByTaskIdIn(tasksWithAccess);
+        if(tasksWithAccess.size()>0) {
+            taskRepository.deleteAllById(tasksWithAccess);
+            shareRepository.deleteByTaskIdIn(tasksWithAccess);
+            return tasksWithAccess;
+        }
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No tasks deleted");
     }
 
 //    @Operation(security = {@SecurityRequirement(name = "Authorization")})
